refactor GameHandler trait to `runtime.rs`
