+++ src/engine.rs
@@ -1,3 +1,2 @@
 pub mod ecs;
-pub mod game;
 pub mod view;
